requirements.txt
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
dash==4.0.0
Flask==3.1.3
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jplephem==2.24
MarkupSafe==3.0.3
narwhals==2.16.0
nest-asyncio==1.6.0
numpy==2.4.1
packaging==26.0
pandas==3.0.0
pillow==12.1.1
plotly==6.5.2
pyarrow==23.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
retrying==1.4.2
scipy==1.17.0
setuptools==82.0.0
sgp4==2.25
six==1.17.0
skyfield==1.54
SQLAlchemy==2.0.46
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
zipp==3.23.0
